Added new `snapshot init` code
