+++ lib/assets/SnapshotHelper.swift
